[ir] Use TINT_IR_* macros in transforms
Convert all macros that produce ICEs to the new ICE macros, with the
exception of some utility functions that do not already have access to
the IR module.

The only macro not changed is TINT_ICE_ON_NO_MATCH, which needs extra
work to add the callback.

Bug: 434966794
Change-Id: I2be4fdd572d0073952a88ae31357946f6f9094f5
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264279
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>