Fix short-circuit override handling.
When generating a short circuit into the IR we need to validate that
both sides are override or const expressions. This fixes an issue where
we'd incorrectly translate an override with the expression in the RHS
into an `if` instead of a `constexpr_if`.

Fixed: 447842134
Change-Id: If76654c92ae61bfa6f501af140f58f99b325f689
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264256
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>