Make bgra8unorm storage read-only an error
The spec says this is an error. Dawn accidentally allowed it.
We switched it to deprecated. We see no usage in the wild so
it seems okay to make it an error like it is supposed to be.

Bug: 427681156,430635197
Fixes: 430635197
Change-Id: Ide5b1251d0691946e47f243d9f1cf8aab13422f4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264095
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Auto-Submit: Gregg Tavares <gman@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>