[dawn][native] Remove the ExternalTextureExpansionsMap
The last use was in ValidateCompatibilityOfSingleBindingWithLayout
to know if a binding was an ExternalTexture. Instead handle external
textures there as any other kind of bindings.

This makes ExternalTexture correctly validated that they are in the
visibility of the BGL, which wasn't the case previously (so a validation
is added).

As a drive-by the validation of static samplers is also fixed to check
for the visibility in the BGL and that the comparisoness in the shader
matches the BGL.

Bug: 42240282
Change-Id: I08079a5a6351850c27313e6f69a5b10e73f2623f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264034
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Geoff Lang <geofflang@chromium.org>