Explicitly list Node architectures
Updates Dawn's NodeJS dependency to explicitly fetch CIPD packages for
all supported architectures for an OS instead of relying on
${{platform}}. Users of this dependency are also updated to select the
correct copy based on the OS/arch combination at runtime.

This is necessary in order to properly support running Node-based tests
on OSes that support multiple architectures. Using ${{platform}} when
fetching dependencies implicitly assumes that the architecture of the
machine fetching dependencies/building is the same as the machine that
will run tests. This assumption does not hold true now that we support
running tests on Swarming, as Mac in particular prefers to build on
arm64 regardless of which architecture is being built for.

Bug: 441328362
Change-Id: Ida2786ff421e35effc812d40a83af951ef3dbb48
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264634
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>