dawn_native GN: copy FXC DLL when building d3d11/12 in standalone
With Rafael's recent changes, we now expect to load the FXC DLL from the
output directory, not the system one, so we make sure to copy it when
building dawn_native.

Fixes dawn.node failing to load FXC when, for e.g., running with d3d11
backend.

Bug: 448635122
Change-Id: I53841668be14b622efa7a5d0e5308eb5571c4f9f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264734
Reviewed-by: Rafael Cintron <rafael.cintron@microsoft.com>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Commit-Queue: Antonio Maiorano <amaiorano@google.com>