[dawn][native] Validate that BGEntry has a single chained structure.
Previously only ExternalTextureBindingEntry was possible to chain in
BindGroupEntry but with the introduction of TexelBufferBindingEntry
there are two types that can be chained.

We need to validate that only one of the two types are allowed. This CL
adds the validation inside the respective Validate[*]Binding functions.

Adds tests for this both for external texture being chained in a texel
buffer and a texel buffer being chained in an external texture.

Bug:
Change-Id: I86fbe10fd193a3978519628385afe36e7cc43927
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262214
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>