[dawn][native] Remove BGL::GetBindingIndex(BindingNumber)
BindingNumbers only make sense to convert to APIBindingIndex since
that's the integer type used to represent packed bindings that match
user-facing bindings.

One use of GetBindingIndex in the GL backend is only for sampled
textures and can be replaced with AsBindingIndex(GetAPIBindingIndex).

The other use of GetBindingIndex was in the Vulkan backend to convert
the BindingNumber in StaticSamplerBindingInfo::sampledTextureBinding to
a BindingIndex. In this CL that member is replaced with a BindingIndex
in the same BGL, which requires updating the StaticSamplerBindingInfo
after the reordering of bindings in the BGL creation, and updates to the
BindGroup validation code.

Bug: 42240282
Change-Id: I0eabaabcea69f3aeae18753c160890af1419b663
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264035
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>