[metal] Expose the exact subgroup size for Apple GPUs
Apple GPUs have a fixed subgroup size of 32. Exposing the actual
subgroup size makes it easier to write high performing shaders,
particularly when using the subgroup matrix extension which requires
the workgroup size to be a multiple of the maximum subgroup size.

This does not increase the ability to fingerprint these devices, since
the presence of the subgroup matrix extension on Metal is already an
indication that the device is an Apple GPU.

Fixed: 416313648
Change-Id: Iaa701d290cae2531a77b14c83b5b7e6d375f6de0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264274
Reviewed-by: Peter McNeeley <petermcneeley@google.com>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>