[tint] Fix StreamStream::Reset()
This was not actually clearing the contents of the underlying
std::stringstream. This caused the copy-assignment implementation to
append existing contents instead of overwriting it, which led to
strange output for switch-case statements in the disassembler and
hardening errors due to a sort that relied on case selector names.

Change-Id: I90e04e0c5218424e6ae2aa03e7524006d53d3c73
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/265214
Auto-Submit: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>