Explicitly list Go architectures
Updates Dawn's Go dependency to explicitly fetch CIPD packages for all
supported architectures for an OS instead of relying on ${{platform}}.
Known users of this dependency are also updated to select the correct
copy based on the OS/arch combination at runtime. However, since the old
files are not automatically removed, it is possible that things may
break at a later point when the builder cache is cleared if a user of Go
was missed.

This is necessary in order to properly support running tests which use
the Go code under //tools/src on OSes that support multiple
architectures. Using ${{platform}} when fetching dependencies implicitly
assumes that the architecture of the machine fetching dependencies is
the same as the machine that will run tests. This assumption does not
hold true now that we support running tests on Swarming, as Mac in
particular prefers to build on arm64 regardless of which architecture is
being built for.

Bug: 441328362
Change-Id: I8829b17f31d7604e1e494e99cc50fd257484c9e1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264976
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>