Validate padding size.
In order to fix issues in backends, limit the padding size allowed in
structures. The GLSL backend ends up emitting a structure member for
each set of padding bytes, when there is megabytes of padding, this is a
lot of members to emit and causes timeouts.

As a first pass, allow 10meg of padding to be inserted as a maximum
between structure members.

Fixed: 415096334
Change-Id: I5d18dbaa2bb11b3861a170ef114492668da7ee8f
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264076
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>