Allow ErrorScope resolution to block on AsyncTasks
Store a list of pending async tasks in error scopes. These tasks are
gathered during PopErrorScope and a WaitListEvent is used to wait until
they are all complete before the PopErrorScope future is marked
complete.

It is not always known which types of errors will be generated by an
AsyncTask so it may write to different error scopes. To handle this, the
AsyncTasks are associated with multiple error scopes with a filter on
which type of error should be handled.

Add a new child class ErrorGeneratingAsyncTask to wrap tasks which
return MaybeError.

Bug: 406522796
Change-Id: Id2cfaebcde38b3c9055d9d29fecd68559ec944e4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/259255
Commit-Queue: Geoff Lang <geofflang@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>