Fix intrinsic match for spirv intrinsics.
For the `image` and `sampled_image` intrinsics they are renamed to have
an `op_` prefix to differentiate them from the equivalent types and the
images are moved to be explicit template parameters. This fixes the
matching algorithm failing to match when they're implicit.

Fixed: 448679392
Change-Id: I718952c614071863148ee8669ad7c4c0265eca34
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/265456
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>