[dawn][native] Add operator== for wgpu::TextureComponentSwizzle
Add and use operator== and != for wgpu::TextureComponentSwizzle.

Also remove TODOs about optimizing view creation for swizzles on <4
channel textures because I realized there's probably no point in doing
that. (It's a minor optimization to prevent creating view objects, and
it only applies in situations where the app is creating views with
swizzles that are equivalent to no swizzle, which should be uncommon
since it's not useful for anything.)

Bug: 414312052
Change-Id: I84b63fe4330ecbf6a14272ce6c0645c52aff65f8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/265334
Commit-Queue: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Reviewed-by: François Beaufort <fbeaufort@google.com>