Fix incorrect continue target.
A single-block loop inside the continuing of a parent loop may have the
continuing block as it's continuing target. Previously we'd set the
outer loop as the continue target in the IR which is invalid. Instead,
when we're emitting the continue block, remove the continuing
information so it can be re-added when we emit the inner loop.

Fixed: 42250521
Change-Id: I77c4fe115008eee6c68ed9c20131d89a573d85c4
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/265714
Commit-Queue: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>