Avoid use-after-move in Blob.h
I noticed some interesting crashes [1] when working on an unrelated
task. Gemini pointed out that it's probably not a good idea to hold onto
vec's data pointer after we've moved it. While much of the time the new
vector might take over the underlying pointer, this won't work if the
vector implementation uses Small Vector Optimization (SVO). In that case
vec.data() would be pointing to the stack and then after the move (which
copies the stack data to the heap) the pointer is invalid.

By grabbing the data and size from the moved vector, we'll avoid the
invalid pointer.

[1]
https://logs.chromium.org/logs/skia/73de020340d43911/+/steps/dm/0/stdout

Change-Id: I3548d2c414a43ad0388996466496e1644f3eef67
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/265755
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>