[metal] Implement commands scheduled future
We need a thread-safe way to call waitUntilScheduled for a Dawn device
for DrDC + Graphite on Mac. The existing WaitForCommandsToBeScheduled()
method isn't thread-safe if the implicit_device_synchronization feature
is not enabled for the Dawn device e.g. for WebGPU.

This CL adds a `SharedTextureMemoryMetalEndAccessState` extension struct
that contains a `commandsScheduledFuture`. This future is ready when
the Metal command buffer submitted on `EndAccess` is in scheduled state.
This future can be waited on any thread with wgpu::Instance::WaitAny.

Bug: chromium:444702048
Change-Id: Icd97c05d6274d5cf9f607a0019e1d9ec6a6a6964
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/265038
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Sunny Sachanandani <sunnyps@chromium.org>
Auto-Submit: Sunny Sachanandani <sunnyps@chromium.org>
Reviewed-by: Corentin Wallez <cwallez@chromium.org>