Allow RTTI to be enabled in the cmake build
I ran into a problem when Skia was running a UBSAN build that
linked in Dawn built with CMake. We'd see run-time crashes like
    ..._functional/function.h:377:17: runtime error: member call on address 0x00016bd88420 which does not point to an object of type ...
    note: object has invalid vptr

The root cause was that -fsanitizer=vptr (which is enabled by
default when -fsanitizer=undefined is set) requires RTTI (Run-Time
Type Information) to work [1]. When Skia builds with UBSAN (which
is combined with our ASAN build), it enables rtti [2] for
this reason.

Either Skia needs to disable the vptr sanitizer or Dawn's cmake
build needs to be configurable. This is the latter option and
I also have Dawn feed the enable RTTI part to SPIRV [3] to compile
all of Dawn and its dependencies with RTTI. I tested this
locally with Skia's GN->CMake system [4] and it seems to build
and run. There might be more third_party deps to configure beyond
these, but it was hard for me to tell for sure.

[1] https://clang.llvm.org/docs/UndefinedBehaviorSanitizer.html
```
-fsanitize=vptr: ... Incompatible with -fno-rtti.
[2] https://skia.googlesource.com/skia/+/351920c5c618bc14b97e588942c596f3e9313b0e/gn/skia/BUILD.gn#681
[3] https://github.com/KhronosGroup/SPIRV-Tools/blob/b9997dafc791634ce51f6bed7ab886ea0e600a15/CMakeLists.txt#L48
[4] https://skia-review.googlesource.com/c/skia/+/1037096

Change-Id: Ie56c1efb64f6a8849aeae61f4ec3b33246f83a4d
Bug: b:434977307
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/266114
Auto-Submit: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Peter McNeeley <petermcneeley@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: David Neto <dneto@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>