[WebGPU backend] Turn on as many passed end2end tests as possible
Suppress and lightly triaged failing reason to fix later.

Add "webgpu" as a valid arg of --backend flag in dawn test.

Add an extension AdapterPropertiesWGPU to chain in adapterInfo
to get the inner backend type of a WebGPUBackend so that
the dawn test can better filter when suppressing.

Bug: 413053623
Change-Id: I0c2f018dda6e04f1fde3e6af7abdf0baa910544d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/263616
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Shrek Shao <shrekshao@google.com>