[dawn][native] Factor the logic to compute the tint::Bindings
This option passed to Tint compilation is used to remap bindings from
WGSL bind points, to whatever the backend needs to match how it sets
BindGroups. The logic in all backends look extremely similar apart from
how they map a (group, index) to a tint::BindPoint.

The logic is extracted in a GenerateBindingRemapping helper function in
TintUtils that takes a backend-specified functor for the mapping of a
(group, index) to a tint::BindPoint.

In the Vulkan backend, some leftover logic for static samplers is
extracted from the generation of the tint::Bindings to modify it a
posteriori.

Bug: 447157265
Change-Id: Ib128c83e71a4ef1916fbe3e1c9475b5f1b680566
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/266375
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>