[dawn][native] Validate that only pinned usages are allowed.
When a texture is pinned it cannot be used as any other usage as it
would break the invariant that pinned textures have the correct memory
barriers done without any extra work. This validation was previously
left as a TODO.

Adds tests and also removes the ASSERT in the Vulkan backend that was
trying to catch incorrect use of pinned textures.

Bug: 435317394
Change-Id: I042b981506690e34908dffb822f8a790d7f7b5f1
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/266614
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>