[spirv] Use DecomposeUniformAccess instead of Std140
This will allow us to remove the requirement for
`uniformAndStorageBuffer16BitAccess` to enable f16, and prepares the
backend to relax uniform layout restrictions.

Add a special-case for single-element swizzles in the printer, since
the `DecomposeUniformAccess` transform does this.

Round uniform buffer sizes up to a multiple of 16 bytes in the Vulkan
backend to support the vec4u polyfill.

Suppress a CTS regression that appears to be a pre-existing Mesa bug.

Suppress a whitebox Vulkan test on Mali due to an flaky failure caused
by the rounding up of buffer sizes.

Put this change behind a feature flag (enabled by default), so that we
can killswitch it if something goes wrong in production. Add a toggle
as well so that we can test the fallback path and so that we can
detect the feature for relaxing f16 requirements.

Bug: 448452698
Fixed: 447407471
Include-Ci-Only-Tests: True
Change-Id: Ia54e9295fe745f702d31ca59d4edc0873256367c
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/264415
Reviewed-by: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>