Fix parsing of a semicolon before a break-if.
When parsing a `continuing` statement we handle the `break-if` parsing
specially so it doesn't collide with the general `break` parsing. To do
so, we check explicitly and then if it doesn't match we hand off to the
statement parsing.

In the statement parsing a `;` by itself is ignored. This means we can
accidentally jump over the `break-if` parsing with a `;`, land in the
statement parsing and mistake the `break-if` for a break and error that
the `break` is missing a `;`.

Fix this by skipping `;` in the continuing code before checking if we
have a `break-if`.

Fixed: 453961764
Change-Id: I02b209d9994df17d5f65e8432269473bd2ef60ea
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/268038
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>
Reviewed-by: James Price <jrprice@google.com>