Enable Win/x64/GTX 1660 tests
Enables tests on the Win/x64/NVIDIA GTX 1660 tester that are equivalent
to the ones run on the Chromium equivalent.

As a result, sharding is increased on all platforms to account for
slower tests on Windows without needing to use test_suite_exceptions.pyl
and the existing test bundle is renamed to include "common" since there
are some Windows-specific variants of the tests.

Bug: 452406604, 454365243
Change-Id: Ib0a3517e4f15ca986501ba550d17639020927bee
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/268263
Commit-Queue: Yuly Novikov <ynovikov@chromium.org>
Auto-Submit: Brian Sheedy <bsheedy@google.com>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>