[tint][uniformity] ignores post-loop code if loop only returns
Consider:

  loop {
    return;
    continuing { break if true; }
  }
  post_loop_code();

The loop behaviour is exactly {Return}, and so uniformity issues
in post_loop_code() should be ignored.

The short-circuiting of the analysis applies to all kinds of loops:
loop, for, and while.

While loops always have conditions, and so the behaviour of a while
loop always includes Next.

Similarly, a for loop with a non-empty condition clause will
have a behaviour including Next.

Fixed: 452692768
Change-Id: I6a6a6964630ea42948e308d037dbe75524289c5d
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/267474
Commit-Queue: David Neto <dneto@google.com>
Reviewed-by: James Price <jrprice@google.com>