[null] Add a simple Null backend to Tint.
This CL adds a `Null` backend which just runs SingleEntryPoint and
SubstituteOverrides and then returns the resulting workgroup
information. It is used by the `DeviceNull` backend in Dawn removing
the need to call the transforms directly. This makes `DeviceNull` behave
in a fashion similar to the other devices.

Change-Id: I80ca8ac1831f877198ec5ab51cfed6c2041f36b3
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/268257
Reviewed-by: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>