[hlsl] Polyfill subgroup_id and num_subgroups
If the workgroup size is linear, we can determine these values from
the location invocation index, the subgroup size, and the workgroup
size.

Otherwise, use an atomic workgroup counter that is incremented by each
subgroup to determine the the ID, and then broadcast that to every
other invocation in the subgroup. The number of subgroups is then the
final value in that counter.

Bug: 454652006
Change-Id: Ide6e51830b93674f1a6c15951244f8e0aebbb32e
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/268794
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>