Polyfill unpack2x16{su}norm on Apple M3+ devices.
The unpack2x16{su}norm calls on M3+ devices appear to be returning the
wrong value when the source value comes from a non-4-byte aligned value.
It appears to be taking the original alignment of the word and using
that instead of the shifted values we're creating.

Fixed: 449576833
Change-Id: Iceeb2970eaad6aa969ace3c4db0a75187fa5a607
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/270054
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: James Price <jrprice@google.com>
Auto-Submit: dan sinclair <dsinclair@chromium.org>