[tint][benchmark] Use TINT_ASSERT instead of SkipWithError
SkipWithError does not result in a non-zero exit code. The benchmarks
should never fail to run, so we should just crash instead if these
conditions fail. This will allow us to add a benchmark run to our CQ
to ensure that we do not break the benchmarks again.

Bug: 445456830
Change-Id: Iee42a87d795d1f157fee7bf3461f74de8fe11d97
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/269834
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>