Fix WebGPU-On-WebGPU use after free issue
The previous solution was associating pipelines
with BindGroupLayoutInternals. That was just flat
out invalid as BindGroupLayoutInternals are deduped
so the association was incorrect.

The association was needed for bindGroupLayouts
that are implicitly created by pipelines using
automatic layout as we need to know which pipeline
to create to get the implicitly generated
bindGroupLayout.

The new solution gathers all the pipelines used in
a command buffer first and serialized them. That way,
ids for the implicit bindGroupLayouts already exist
and it's safe to then serialize the bindGroups.

Bug: 455605671
Change-Id: I78e866eda372df48f37e0b6b91e8d6c76199e458
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/269574
Commit-Queue: Gregg Tavares <gman@chromium.org>
Reviewed-by: Loko Kung <lokokung@google.com>