[ir] Move type member to Value base class
All but one subclass of Value declared its own type member and
SetType() method, as well as overriding the virtual Type()
getter. Instead, just declare the type in the base class and define
those methods in the base class too.

As well as removing an unnecessary virtual method and duplicated code
in the subclasses, this also makes it possible to call SetType() on a
value without first casting it to a subclass, which is used to
slightly simplify a callsite in ChangeImmediateToUniform.

Change-Id: Ib1f34b01fef841071e02a868a216b76cfe517f80
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/271117
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>