[msl] Fixup subgroup matrix initialization.
When creating a `subgroup_matrix` in WGSL when provided with an argument
will create a filled matrix. The MSL behaviour when calling the type
constructor is to make a diagonal matrix. Switch to using
`make_filled_simdgroup_matrix` instead to match the expected WGSL
behaviour.

An empty construct call is converted to a make filled simdgroup matrix
with the appropriate 0 value to make sure the memory is initialized.

Fixed: 457816671
Change-Id: I48ffd3a5ca6b5f0d88d8900d469e8814af1170ef
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/271414
Auto-Submit: dan sinclair <dsinclair@chromium.org>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Reviewed-by: James Price <jrprice@google.com>