Fix BufferBase::mMapData race
BufferBase::mMapData is accessed from UnmapInternal() when PendingMap
and it's assumed that it contains Ref<MapAsyncEvent>. The variable is
also written from BufferBase::FinalizeMap(). There is no mutex that
protects against reading in UnmapInternal() while writing in
FinalizeMap() which is problematic.

Fix this by protecting both places Ref<MapAsyncEvent> is reset with the
same mutex. The existing mutex in MapAsyncEvent isn't usable since we
have to access the variable and dereference it to get mutex from
BufferBase. Fix this by making BufferBase support WeakRef and own the
mutex. MapAsyncEvent can promote a WeakRef<BufferBase>, if promotion
fails the buffer has been destroyed and there is no possible race. If
promotion is successful then MapAsyncEvent is able to lock mutex before
proceeding.

Additionally MapAsyncEvent::mStatus is also initially accessed with the
same mutex locked. The status is changed to not be success if outcomes
race and Unmap() path wins, so if MapAsyncEvent::Complete() sees a
non-success status it knows map was aborted and doesn't need to bother
checking mPendingMapEvent.

mMapdata is split into mMappedPointer and mPendingMapEvent to make code
easier to understand.

Bug: 425472913
Change-Id: Iff75404097037eb23cbdfc633ef48a2dfbd165ed
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/269714
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Kyle Charbonneau <kylechar@google.com>