[dawn][node] Add support for ExternalTexture BGLEntry.
Previously only GPUExternalTexture could be bound to a an external
texture entry and GPUExternalTexture was not implemented so we left BGL
support TODO in dawn.node.

Now that GPUTextureView can be used for an external texture BGL entry,
implement BGL support for external textures to pass tests.

Bug:
Change-Id: I2a561fdd722b25d4984258b23aa1ef1b1be657f2
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/272054
Reviewed-by: Antonio Maiorano <amaiorano@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>