Add more directories checked by the clang unsafe buffers plugin.
This change adds tint, dawn/samples and dawn/utils to the directories
checked by the unsafe buffers plugin.

Tint's cmake build does have the plugin enabled, but it is not as robust
as the one pulled in from //build as it does not check for usage of
methods such as memcpy.

Bug: 438937152
Change-Id: I49af64eda9316729550cabcbc0582f2ba690c2f0
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/262076
Reviewed-by: James Price <jrprice@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Reviewed-by: Ken Russell <kbr@google.com>
Commit-Queue: Liza Burakova <liza@chromium.org>