[tint][val] Rework how `@builtin(...)` is validated
Refactors the existing checker implementation to support the special
cased built-ins like @builtin(position), so that their special
handling can be removed. This refactoring changes the error logging to
be explicit about what the builtin was attached to, e.g. a vertex
shader input, and provides what the valid options are.

This CL also changes the capability to allow clip_distance on scalars
to also allow vectors, since that is needed by the HLSL
backend. Trying to split them into two capabilities leads to a bunch
of duplicated instances of them and more complex logic in the
validator. (The original validation logic had an oversight that
allowed this construct to pass unintentionally)

Issue: 455376684
Change-Id: Ia5dac3782158c2e1acc3d0505fa7fdf4a1057f3b
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/271854
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>