[spirv] Polyfill 8-bit subgroup matrix construct
SPIR-V requires that the value passed to OpCompositeConstruct has the
same type as the matrix, so we need to convert it to an 8-bit
value. This was already done for construct instructions generated by
the matrix-scalar builtin polyfills.

Update the validator to allow either the 8-bit or 32-bit types for an
8-bit subgroup matrix constructor, to satisfy both WGSL and SPIR-V.

Fixed: 459800105
Change-Id: I46626a3c465bd8394458878862d0ebd3961a52a8
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/272875
Reviewed-by: dan sinclair <dsinclair@chromium.org>
Commit-Queue: James Price <jrprice@google.com>