[spirv] Make all access chain indices unsigned.
For any `OpAccessChain` in SPIR-V it is a validation error to have a
negative constant index value. In order to make sure we never emit
negative indices, always bitcast the index to an unsigned value before
passing to the access chain.

There are currently only two ways this can happen, by disabling
robustness or but the disabling of runtime array robustness. In the case
of the former, robustness has been disabled, so it's on the caller to
make sure the index was never negative. In the case of the latter we
only do this when we have RobustBufferAccess2 enabled and it will
handling the clamping for us.

Fixed: 459509275
Change-Id: I784cb3c8dbc3e3c907402094940fabed16138827
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/273294
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>