[tint] Polyfill 'position' for vulkan in muliti-sampling
The WebGPU/WGSL SPEC specifies that 'position' be pixel centered.
For MSL/HLSL this will always be the case even with multi-sampled
fragments. However vulkan does not appear to always behave this way [0].
This change polyfills 'position' in the case where we require
known consistent behavior.

This change allocates an additional variable (vec4f) for the
interpolant that is introduced to emulate the 'position' builtin.

For a few select devices (intel) this means we need to drop the limit
maxInterStageShaderVariables from 28 to 16.

[0]
https://docs.google.com/document/d/1bEvj-7nSeynmywCUnSetpp06jcIeUjmmFBeOo8-pJLg/edit?pli=1&tab=t.0

Bug: 445242488
Change-Id: If5094d223c7d4e5d654886da9cf838e8623df187
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/269074
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: Peter McNeeley <petermcneeley@google.com>