[dawn][native] Make more EncodingContext logic private.
It was not used outside of EncodingContext but was in the public: block
of the class. Also use DAWN_TRY in RenderPassEncoder to avoid directly
consuming an error in a TryEncode block.

This is in preparation for changes error handling of encoders to match
the latest specification.

Bug: 449980070
Change-Id: I8634037f7f0b6e0133df1ce6270221a64d9d3539
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/272038
Reviewed-by: Loko Kung <lokokung@google.com>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>