[dawn][native] Separate deferErrors to a boolean in EncodingContext
The logic in EncodingContext::HandleError decides whether the error
should be deferred until Finish() or if it should be immediately
forwarded to the device. Future CLs will change that logic so to make
them more clear, separate the condition as a boolean in advance.

This is in preparation for changes error handling of encoders to match
the latest specification.

Bug: 449980070
Change-Id: I0d4b44a4f662b1e5246d496fa362e27c7b34ad77
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/273114
Reviewed-by: Loko Kung <lokokung@google.com>
Reviewed-by: Kai Ninomiya <kainino@chromium.org>
Commit-Queue: Corentin Wallez <cwallez@chromium.org>