[tint][val] Proof-of-concept for reworking IOAttribute validation
Use a more data/table driven pattern for IOAttribute validation, akin
to BuiltInValidation.

Invariant validation is reimplemented using this new framework as a
proof-of-concept, other IOAttributes will be migrated in follow on
CLs.

An enum and ToString for tagging IOAttributes is added to allow baking
into the checker what attribute is being validated.

There is also some minor clean-ups to related code
included. Specifically for the called impl in WalkTypeAndMembers the
context param is moved to the beginning of the param list to match the
form as other calls, and CheckVertexEntryPoint is renamed to
CheckPositionPresentForVertexOutput, since that is all it does now.

Issue: 455376684

Change-Id: I791210b7d1d12ff99879fff55223592803e0504a
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/273175
Reviewed-by: James Price <jrprice@google.com>
Commit-Queue: dan sinclair <dsinclair@chromium.org>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: dan sinclair <dsinclair@chromium.org>